+++ dydx/dydxPresenters/dydxPresenters/_v4/TakeProfitStopLoss/dydxTakeProfitStopLossViewPresenter.swift
@@ -43,7 +43,8 @@ private protocol dydxTakeProfitStopLossViewPresenterProtocol: HostedViewPresente
 
 private class dydxTakeProfitStopLossViewPresenter: HostedViewPresenter<dydxTakeProfitStopLossViewModel>, dydxTakeProfitStopLossViewPresenterProtocol {
     fileprivate var marketId: String?
-
+    @SynchronizedLock private var pendingOrders: Int?
+    
     deinit {
         clearTriggersInput()
     }
@@ -179,13 +180,13 @@ private class dydxTakeProfitStopLossViewPresenter: HostedViewPresenter<dydxTakeP
 
         let formattedTakeProfitUsdcDiff = dydxFormatter.shared.raw(number: triggerOrdersInput?.takeProfitOrder?.price?.usdcDiff?.doubleValue, digits: 2) ?? ""
         let formattedTakeProfitUsdcPercentage = dydxFormatter.shared.raw(number: triggerOrdersInput?.takeProfitOrder?.price?.percentDiff?.doubleValue, digits: 2) ?? ""
-        viewModel?.takeProfitStopLossInputAreaViewModel?.gainInputViewModel?.programmaticallySet(value: formattedTakeProfitUsdcDiff, forUnit: .dollars)
-        viewModel?.takeProfitStopLossInputAreaViewModel?.gainInputViewModel?.programmaticallySet(value: formattedTakeProfitUsdcPercentage, forUnit: .percentage)
+        viewModel?.takeProfitStopLossInputAreaViewModel?.gainInputViewModel?.dollars = formattedTakeProfitUsdcDiff
+        viewModel?.takeProfitStopLossInputAreaViewModel?.gainInputViewModel?.percentage = formattedTakeProfitUsdcPercentage
 
         let formattedStopLossUsdcDiff = dydxFormatter.shared.raw(number: triggerOrdersInput?.stopLossOrder?.price?.usdcDiff?.doubleValue, digits: 2) ?? ""
         let formattedStopLossUsdcPercentage = dydxFormatter.shared.raw(number: triggerOrdersInput?.stopLossOrder?.price?.percentDiff?.doubleValue, digits: 2) ?? ""
-        viewModel?.takeProfitStopLossInputAreaViewModel?.lossInputViewModel?.programmaticallySet(value: formattedStopLossUsdcDiff, forUnit: .dollars)
-        viewModel?.takeProfitStopLossInputAreaViewModel?.lossInputViewModel?.programmaticallySet(value: formattedStopLossUsdcPercentage, forUnit: .percentage)
+        viewModel?.takeProfitStopLossInputAreaViewModel?.lossInputViewModel?.dollars = formattedStopLossUsdcDiff
+        viewModel?.takeProfitStopLossInputAreaViewModel?.lossInputViewModel?.percentage = formattedStopLossUsdcPercentage
 
         // logic primarily to pre-populate custom amount. need to account 3 situations: 1 take profit order or 1 stop loss order or both
         if let customSize = triggerOrdersInput?.size?.doubleValue.magnitude, customSize != position?.size?.current?.doubleValue.magnitude {
@@ -321,24 +322,7 @@ private class dydxTakeProfitStopLossViewPresenter: HostedViewPresenter<dydxTakeP
         viewModel.takeProfitStopLossInputAreaViewModel?.takeProfitPriceInputViewModel = .init(title: DataLocalizer.shared?.localize(path: "APP.TRIGGERS_MODAL.TP_PRICE", params: nil))
         viewModel.takeProfitStopLossInputAreaViewModel?.stopLossPriceInputViewModel = .init(title: DataLocalizer.shared?.localize(path: "APP.TRIGGERS_MODAL.SL_PRICE", params: nil))
 
-        viewModel.takeProfitStopLossInputAreaViewModel?.gainInputViewModel = .init(triggerType: .takeProfit) { (value, unit) in
-            switch unit {
-            case .dollars:
-                AbacusStateManager.shared.triggerOrders(input: value, type: .takeprofitusdcdiff)
-            case .percentage:
-                AbacusStateManager.shared.triggerOrders(input: value, type: .takeprofitpercentdiff)
-            }
-        }
-        viewModel.takeProfitStopLossInputAreaViewModel?.lossInputViewModel = .init(triggerType: .stopLoss) { (value, unit) in
-            switch unit {
-            case .dollars:
-                AbacusStateManager.shared.triggerOrders(input: value, type: .stoplossusdcdiff)
-            case .percentage:
-                AbacusStateManager.shared.triggerOrders(input: value, type: .stoplosspercentdiff)
-            }
-        }
-
-#if DEBUG
+        #if DEBUG
         viewModel.shouldDisplayCustomLimitPriceViewModel = AbacusStateManager.shared.environment?.featureFlags.isSlTpLimitOrdersEnabled == true
         #endif
 
@@ -365,6 +349,22 @@ private class dydxTakeProfitStopLossViewPresenter: HostedViewPresenter<dydxTakeP
         }
         viewModel.customLimitPriceViewModel?.stopLossPriceInputViewModel?.onEdited = {
             AbacusStateManager.shared.triggerOrders(input: $0, type: .stoplosslimitprice)
+        }
+        viewModel.takeProfitStopLossInputAreaViewModel?.gainInputViewModel = .init(triggerType: .takeProfit) { (value, unit) in
+            switch unit {
+            case .dollars:
+                AbacusStateManager.shared.triggerOrders(input: value, type: .takeprofitusdcdiff)
+            case .percentage:
+                AbacusStateManager.shared.triggerOrders(input: value, type: .takeprofitpercentdiff)
+            }
+        }
+        viewModel.takeProfitStopLossInputAreaViewModel?.lossInputViewModel = .init(triggerType: .stopLoss) { (value, unit) in
+            switch unit {
+            case .dollars:
+                AbacusStateManager.shared.triggerOrders(input: value, type: .stoplossusdcdiff)
+            case .percentage:
+                AbacusStateManager.shared.triggerOrders(input: value, type: .stoplosspercentdiff)
+            }
         }
 
         // set up toggle interactions
@@ -406,9 +406,6 @@ private class dydxTakeProfitStopLossViewPresenter: HostedViewPresenter<dydxTakeP
 
         self.viewModel = viewModel
     }
-
-    @SynchronizedLock var pendingOrders: Int?
-
 }
 
 private extension Abacus.OrderSide {
@@ -418,32 +415,5 @@ private extension Abacus.OrderSide {
         case .sell: return .long_
         default: return .short_
         }
-    }
-}
-
-import Darwin
-
-@propertyWrapper
-public struct SynchronizedLock<Value> {
-    private var value: Value
-    private var lock = NSLock()
-
-    public var wrappedValue: Value {
-        get { lock.synchronized { value } }
-        set { lock.synchronized { value = newValue } }
-    }
-
-    public init(wrappedValue value: Value) {
-        self.value = value
-    }
-}
-
-private extension NSLock {
-
-    @discardableResult
-    func synchronized<T>(_ block: () -> T) -> T {
-        lock()
-        defer { unlock() }
-        return block()
     }
 }
