+++ dydx/dydxStateManager/dydxStateManager/AbacusStateManager.swift
@@ -428,7 +428,9 @@ extension AbacusStateManager {
         case success
         case failed(Abacus.ParsingError?)
     }
-
+    
+    /// places the currently drafted trigger order(s)
+    /// - Returns: the number of resulting cancel orders + place order requests
     public func placeTriggerOrders(callback: @escaping ((SubmissionStatus) -> Void)) -> Int? {
         let payload = asyncStateManager.commitTriggerOrders { successful, error, _ in
             if successful.boolValue {
